requirements.txt
python-telegram-bot
requests
toml
peewee
beautifulsoup4
